fix: app lifecycle, phantom hello, polish pipeline, and stability improvements
App lifecycle:
- Add NSApplicationDelegate with applicationWillTerminate to cleanly stop
  backend on quit
- PID file (/tmp/voxflow-backend.pid) survives crashes — stale backend
  recovered on relaunch
- atexit() fallback for last-resort backend cleanup
- Add "Quit VoxFlow" menu item with Cmd+Q

Phantom "hello" insertion fix:
- Add silence detection (RMS energy gate) in stopAndValidateAudio — discards
  silent audio before transcription runs, preventing phantom Fn key triggers
  from producing hallucinated text
- Move hello/hi/hey to always-filtered in both Swift and Python hallucination
  filters (was short-audio-only, letting 3s+ silence through)
- Sync Python hallucination list with Swift (was missing hello/hi/hey entirely)
- Add low-confidence gate (<0.15) for short text (<=2 words) before auto-insert

Backend robustness:
- lsof now filters -sTCP:LISTEN (was matching active connections)
- WhisperEngine/PolishEngine use _load_failed flag instead of _pipeline=False
  sentinel — allows retry after transient failures
- model_loaded property no longer triggers model loading as side effect
- Crash restart exhaustion (3x) now surfaces reason to UI
- uvicorn gets --timeout-graceful-shutdown 4

Polish pipeline:
- Better prompt with behavioral tone instructions for flan-t5-small
- Dynamic max_new_tokens scaled to input length
- Guardrail skips length ratio for short inputs (<=5 words)
- Wider length tolerance for medium inputs (6-10 words)
- Improved echo detection (normalized comparison)

Performance:
- FocusContextMonitor AX polling moved to background DispatchSourceTimer
- Paste delays reduced (150ms→100ms activation, 50ms→30ms Electron)

Other fixes:
- WhisperKit latency calculation (was off by 1000x in sub-second component)
- Audio tap cleanup on engine.start() failure
- Readiness badge reflects WhisperKit status when backend is down
- NSAccessibilityUsageDescription added to Info.plist template
- Dead PRODUCT_BIN line removed from build script

Tests: 234 Swift + 240 Python — all passing.

+++ Sources/VoxFlowApp/Services/WhisperKitSTTService.swift
@@ -52,7 +52,7 @@ final class WhisperKitSTTService {
 
         let text = results.map(\.text).joined(separator: " ").trimmingCharacters(in: .whitespacesAndNewlines)
         let elapsed = started.duration(to: .now)
-        let latencyMs = Int(elapsed.components.seconds * 1000 + elapsed.components.attoseconds / 1_000_000_000_000_000)
+        let latencyMs = Int(elapsed.components.seconds) * 1000 + Int(elapsed.components.attoseconds / 1_000_000_000_000_000)
 
         let confidence: Double = {
             guard let first = results.first, let seg = first.segments.first else { return 0.0 }
